+++ lib/uplinkc/testdata/bucket_test.c
@@ -8,22 +8,22 @@
 #include "uplink.h"
 #include "helpers2.h"
 
-void HandleProject(Project project);
+void handle_project(Project project);
 
 int main(int argc, char *argv[]) {
-    WithTestProject(&HandleProject);
+    with_test_project(&handle_project);
 }
 
-void HandleProject(Project project) {
+void handle_project(Project project) {
     char *_err = "";
     char **err = &_err;
 
     char *bucket_name = "TestBucket";
 
-    BucketConfig config = TestBucketConfig();
+    BucketConfig config = test_bucket_config();
     BucketInfo info = create_bucket(project, bucket_name, &config, err);
     require_noerror(*err);
-    FreeBucketInfo(&info);
+    free_bucket_info(&info);
 
     EncryptionAccess access = {};
     memcpy(&access.key[0], "hello", 5);
@@ -38,6 +38,6 @@ void HandleProject(Project project) {
         //    
         //}
     }
-    CloseBucket(bucket, err);
+    close_bucket(bucket, err);
     require_noerror(*err);
 }
